requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
googlesearch-python==1.2.3
googletrans==4.0.0rc1
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
protobuf==4.25.1